Test split into own files. Coverage now 100%
