+++ jest.config.js
@@ -14,10 +14,10 @@ module.exports = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 50,
-      functions: 50,
-      lines: 50,
-      statements: 50
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100
     }
   },
