+++ src/min/index.ts
@@ -1,16 +1,20 @@
-interface Array<T> {
-  /**
-   * Shorthand for finding element with minimum value of an attribute.
-   *
-   * @returns The entire element found. So if you write `x => x.score` x is returned.
-   *
-   * Examples:
-   * ```typescript
-   * const minScore = myArr.min(x => x.score);
-   * const minScore = [1,2,3].min();
-   * ```
-   */
-  min(callbackfn?: (value: T) => number, thisArg?: any[]): T;
+import "../sortByAttr";
+
+declare global {
+  interface Array<T> {
+    /**
+     * Shorthand for finding element with minimum value of an attribute.
+     *
+     * @returns The entire element found. So if you write `x => x.score` x is returned.
+     *
+     * Examples:
+     * ```typescript
+     * const minScore = myArr.min(x => x.score);
+     * const minScore = [1,2,3].min();
+     * ```
+     */
+    min(callbackfn?: (value: T) => number, thisArg?: any[]): T;
+  }
 }
 
 Array.prototype.min = function (func = x => x, thisArg = this) {
